+++ src/main/java/domain/card/Cards.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,11 +15,11 @@ public class Cards {
     public static Cards createDefaultCards() {
         List<Card> defaultCards = new ArrayList<>();
 
-        int[] numbers = new int[]{1, 2, 3, 4, 5, 6, 7, 8, 9};
-        String[] shapes = new String[]{"클로버", "스페이드", "하트", "다이아몬드"};
-        for (String shape : shapes) {
-            for (int num : numbers) {
-                defaultCards.add(new Card(num, shape));
+        Shape[] shapes = Shape.values();
+        Value[] values = Value.values();
+        for (Shape shape : shapes) {
+            for (Value value : values) {
+                defaultCards.add(new Card(value, shape));
             }
         }
         return new Cards(defaultCards);
