+++ src/main/java/service/BlackJackGame.java
@@ -3,8 +3,8 @@ package service;
 import controller.dto.PlayerDeck;
 import controller.dto.PlayerDeckResult;
 import controller.dto.PlayerRevenue;
-import domain.Card;
-import domain.Cards;
+import domain.card.Card;
+import domain.card.Cards;
 import domain.Dealer;
 import domain.Player;
 import domain.Players;
